[InitSystem] Separate InitSystem from policies
Moves `InitSystem` and the related `SystemdInit` subclass out of
`sos/policies/__init__.py` and into a new `sos/policies/init_systems/`
subdir.

Related: #2349
